# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ilios API",
    "description": "Local-first API for Ilios.",
    "version": "0.2.1"
  },
  "paths": {
...
    "/libraries": {
      "get": {
        "tags": [
          "libraries"
        ],
        "summary": "Get Libraries",
        "operationId": "get_libraries_libraries_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Library"
                  },
                  "type": "array",
                  "title": "Response Get Libraries Libraries Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "libraries"
        ],
        "summary": "Post Library",
        "operationId": "post_library_libraries_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Library"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/libraries/{library_id}": {
      "get": {
        "tags": [
          "libraries"
        ],
        "summary": "Get Library By Id",
        "operationId": "get_library_by_id_libraries__library_id__get",
        "parameters": [
          {
            "name": "library_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Library Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Library"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "libraries"
        ],
        "summary": "Delete Library By Id",
        "operationId": "delete_library_by_id_libraries__library_id__delete",
        "parameters": [
          {
            "name": "library_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Library Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryDeleteResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/libraries/{library_id}/archive": {
      "post": {
        "tags": [
          "libraries"
        ],
        "summary": "Archive Library By Id",
        "operationId": "archive_library_by_id_libraries__library_id__archive_post",
        "parameters": [
          {
            "name": "library_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Library Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Library"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/libraries/{library_id}/translations/missing": {
      "post": {
        "tags": [
          "libraries"
        ],
        "summary": "Translate Missing Library Blocks",
        "operationId": "translate_missing_library_blocks_libraries__library_id__translations_missing_post",
        "parameters": [
          {
            "name": "library_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Library Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranslationBatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryTranslationBatchResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Library": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "status": {
            "$ref": "#/components/schemas/LibraryStatus"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "path",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "Library"
      },
      "LibraryCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "path": {
            "type": "string",
            "minLength": 1,
            "title": "Path"
          }
        },
        "type": "object",
        "required": [
          "name",
          "path"
        ],
        "title": "LibraryCreate"
      },
      "LibraryDeleteResult": {
        "properties": {
          "library_id": {
            "type": "string",
            "title": "Library Id"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "deleted_cache": {
            "type": "boolean",
            "title": "Deleted Cache"
          }
        },
        "type": "object",
        "required": [
          "library_id",
          "path",
          "deleted_cache"
        ],
        "title": "LibraryDeleteResult"
      },
      "LibraryStatus": {
        "type": "string",
        "enum": [
          "active",
          "missing",
          "archived"
        ],
        "title": "LibraryStatus"
      },
      "LibraryTranslationBatchResult": {
        "properties": {
          "library_id": {
            "type": "string",
            "title": "Library Id"
          },
          "target_language": {
            "type": "string",
            "title": "Target Language"
          },
          "articles_considered": {
            "type": "integer",
            "title": "Articles Considered",
            "default": 0
          },
          "articles_queued": {
            "type": "integer",
            "title": "Articles Queued",
            "default": 0
          },
          "jobs_created": {
            "type": "integer",
            "title": "Jobs Created",
            "default": 0
          },
          "existing_jobs": {
            "type": "integer",
            "title": "Existing Jobs",
            "default": 0
          },
          "cached_blocks": {
            "type": "integer",
            "title": "Cached Blocks",
            "default": 0
          },
          "skipped_blocks": {
            "type": "integer",
            "title": "Skipped Blocks",
            "default": 0
          },
          "job_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Job Ids"
          },
          "article_results": {
            "items": {
              "$ref": "#/components/schemas/TranslationBatchResult"
            },
            "type": "array",
            "title": "Article Results"
          }
        },
        "type": "object",
        "required": [
          "library_id",
          "target_language"
        ],
        "title": "LibraryTranslationBatchResult"
      },
...
      "TranslationBatchRequest": {
        "properties": {
          "target_language": {
            "type": "string",
            "maxLength": 40,
            "minLength": 2,
            "title": "Target Language",
            "default": "zh-CN"
          },
          "provider_profile_id": {
            "type": "string",
            "minLength": 1,
            "title": "Provider Profile Id"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "glossary_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Glossary Version"
          },
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          },
          "block_uids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Block Uids"
          },
          "custom_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Prompt"
          }
        },
        "type": "object",
        "required": [
          "provider_profile_id"
        ],
        "title": "TranslationBatchRequest"
      },
      "TranslationBatchResult": {
        "properties": {
          "library_id": {
            "type": "string",
            "title": "Library Id"
          },
          "article_revision_id": {
            "type": "string",
            "title": "Article Revision Id"
          },
          "target_language": {
            "type": "string",
            "title": "Target Language"
          },
          "jobs_created": {
            "type": "integer",
            "title": "Jobs Created"
          },
          "existing_jobs": {
            "type": "integer",
            "title": "Existing Jobs",
            "default": 0
          },
          "cached_blocks": {
            "type": "integer",
            "title": "Cached Blocks"
          },
          "skipped_blocks": {
            "type": "integer",
            "title": "Skipped Blocks"
          },
          "job_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Job Ids"
          }
        },
        "type": "object",
        "required": [
          "library_id",
          "article_revision_id",
          "target_language",
          "jobs_created",
          "cached_blocks",
          "skipped_blocks"
        ],
        "title": "TranslationBatchResult"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}